+++ rust/cubestore/cubestore/src/bin/cubestored.rs
@@ -11,13 +11,10 @@ use serde_json::Value;
 use std::collections::HashMap;
 use std::time::Duration;
 use tokio::runtime::Builder;
-use cubestore::cluster::worker_pool::worker_main;
 
 const PACKAGE_JSON: &'static str = std::include_str!("../../../package.json");
 
 fn main() {
-    println!("DDD main {}", std::process::id());
-
     let package_json: Value = serde_json::from_str(PACKAGE_JSON).unwrap();
     let version = package_json
         .get("version")
@@ -37,10 +34,9 @@ fn main() {
     init_metrics("127.0.0.1:0", "127.0.0.1:8125", metrics_mode);
     init_cube_logger(true);
 
-    log::info!("Cube Store version {} {}", version, std::process::id());
+    log::info!("Cube Store version {}", version);
 
     let config = Config::default();
-    let worker_services = Config::configure_worker_services();
 
     let trim_every = config.config_obj().malloc_trim_every_secs();
     if trim_every != 0 {
@@ -52,7 +48,6 @@ fn main() {
 
     #[cfg(not(target_os = "windows"))]
     cubestore::util::respawn::init();
-    println!("DDD init {}", std::process::id());
 
     let mut tokio_builder = Builder::new_multi_thread();
     tokio_builder.enable_all();
@@ -72,7 +67,6 @@ fn main() {
 
         let services = config.cube_services().await;
 
-        println!("DDD config {}", std::process::id());
         track_event("Cube Store Start".to_string(), HashMap::new()).await;
 
         stop_on_ctrl_c(&services).await;