+++ rust/cubestore/cubestore/src/cluster/worker_pool.rs
@@ -1,7 +1,6 @@
 use std::fmt::Debug;
 use std::marker::PhantomData;
 use std::panic;
-use std::backtrace::Backtrace;
 use std::process::Child;
 use std::sync::Arc;
 use std::time::Duration;
@@ -20,10 +19,11 @@ use tokio::sync::{oneshot, watch, Notify, RwLock};
 use tracing::{instrument, Instrument};
 use tracing_futures::WithSubscriber;
 
-use crate::util::catch_unwind::async_try_with_catch_unwind;
-use crate::util::respawn::{IPC_ENV, respawn};
+use crate::config::{Config, WorkerServices};
+use crate::util::respawn::respawn;
 use crate::CubeError;
 use datafusion::cube_ext;
+use datafusion::cube_ext::catch_unwind::async_try_with_catch_unwind;
 
 pub struct WorkerPool<
     T: Debug + Serialize + DeserializeOwned + Sync + Send + 'static,
@@ -52,7 +52,7 @@ pub trait MessageProcessor<
     R: Serialize + DeserializeOwned + Sync + Send + 'static,
 >
 {
-    async fn process(args: T) -> Result<R, CubeError>;
+    async fn process(services: &WorkerServices, args: T) -> Result<R, CubeError>;
 }
 
 impl<
@@ -272,8 +272,6 @@ where
     R: Serialize + DeserializeOwned + Sync + Send + 'static,
     P: MessageProcessor<T, R>,
 {
-    println!("QQQ {:?} {:?}", std::process::id(), std::env::var(IPC_ENV));
-
     let (rx, tx) = (a.args, a.results);
     let mut tokio_builder = Builder::new_multi_thread();
     tokio_builder.enable_all();
@@ -284,16 +282,18 @@ where
     let runtime = tokio_builder.build().unwrap();
     runtime.block_on(async move {
         let config = Config::default();
-        config.configure_worker().await?;
-        // injector is lazy
-        // pull the query executor from injector
+        config.configure_injector().await;
+        let services = config.worker_services().await;
 
         loop {
             let res = rx.recv();
             match res {
                 Ok(args) => {
-                    println!("PPPX{} {:?}", std::process::id(), args);
-                    let result = async_try_with_catch_unwind(P::process(args)).await;
+                    let result =
+                        match async_try_with_catch_unwind(P::process(&services, args)).await {
+                            Ok(result) => result,
+                            Err(panic) => Err(CubeError::from(panic)),
+                        };
                     let send_res = tx.send(result);
                     if let Err(e) = send_res {
                         error!("Worker message send error: {:?}", e);
@@ -322,6 +322,7 @@ mod tests {
     use tokio::runtime::Builder;
 
     use crate::cluster::worker_pool::{worker_main, MessageProcessor, WorkerPool};
+    use crate::config::WorkerServices;
     use crate::queryplanner::serialized_plan::SerializedLogicalPlan;
     use crate::util::respawn;
     use crate::CubeError;
@@ -348,7 +349,7 @@ mod tests {
 
     #[async_trait]
     impl MessageProcessor<Message, Response> for Processor {
-        async fn process(args: Message) -> Result<Response, CubeError> {
+        async fn process(_services: &WorkerServices, args: Message) -> Result<Response, CubeError> {
             match args {
                 Message::Delay(x) => {
                     Delay::new(Duration::from_millis(x)).await;